feat(assignment): add middleware

+++ server/app/Http/Controllers/AssignmentController.php
@@ -6,15 +6,35 @@ use App\Http\Resources\AssignmentResource;
 use App\Models\Course;
 use App\Models\Assignment;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
 
-class AssignmentController extends Controller
+class AssignmentController extends Controller implements HasMiddleware
 {
+
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth:api'),
+            new Middleware('role:teacher'),
+        ];
+    }
+
     /**
      * Display a listing of the resource.
      */
-    public function index(string $course, string $chapter, string $subchapter)
-    {
+    public function index(
+        Request $request,
+        string $course,
+        string $chapter,
+        string $subchapter
+    ) {
         $course = Course::where('slug', $course)->firstOrFail();
+
+        if ($course->user_id !== $request->user()->id) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
         $chapter = $course->chapters()->where('position', $chapter)->firstOrFail();
         $subchapter = $chapter->subchapters()->where('position', $subchapter)->firstOrFail();
 
@@ -24,9 +44,18 @@ class AssignmentController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request, string $course, string $chapter, string $subchapter)
-    {
+    public function store(
+        Request $request,
+        string $course,
+        string $chapter,
+        string $subchapter
+    ) {
         $course = Course::where('slug', $course)->firstOrFail();
+
+        if ($course->user_id !== $request->user()->id) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
         $chapter = $course->chapters->where('position', $chapter)->firstOrFail();
         $subchapter = $chapter->subchapters->where('position', $subchapter)->firstOrFail();
 
@@ -50,17 +79,35 @@ class AssignmentController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(string $course, string $chapter, string $subchapter, Assignment $assignment)
-    {
+    public function show(
+        Request $request,
+        string $course,
+        string $assignment
+    ) {
+        $course = Course::where('slug', $course)->firstOrFail();
+
+        if ($course->user_id !== $request->user()->id) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
+        $assignment = Assignment::findOrFail($assignment);
+
         return new AssignmentResource($assignment);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $course, string $chapter, string $subchapter, Assignment $assignment)
-    {
+    public function update(
+        Request $request,
+        string $course,
+        string $assignment
+    ) {
         $course = Course::where('slug', $course)->firstOrFail();
+
+        if ($course->user_id !== $request->user()->id) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
 
         $validated = $request->validate([
             'title' => 'required|string|max:255',
@@ -68,6 +115,8 @@ class AssignmentController extends Controller
             'due_date' => 'required|date_format:d-m-Y',
             'is_active' => 'required|boolean'
         ]);
+
+        $assignment = Assignment::findOrFail($assignment);
 
         $assignment->fill($validated);
         $assignment->save();
@@ -78,9 +127,19 @@ class AssignmentController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $course, string $chapter, string $subchapter, Assignment $assignment)
-    {
+    public function destroy(
+        Request $request,
+        string $course,
+        string $assignment
+    ) {
         $course = Course::where('slug', $course)->firstOrFail();
+
+        if ($course->user_id !== $request->user()->id) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
+        $assignment = Assignment::findOrFail($assignment);
+
         $assignment->delete();
 
         return response()->json(['message' => 'Assignment have been deleted succesfully'], 204);
